server/requirements.txt
django==5.1.2
celery==5.4.0
beautifulsoup4==4.12.3
redis==5.2.0
django-celery-beat==2.7.0
requests==2.32.3
fake_useragent==1.5.1
lxml==5.3.0
django-dotenv==1.4.2
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.5.0
black==25.1.0
flake8==7.2.0
isort==6.0.1